+++ packages/phone-number-privacy/common/src/interfaces/index.ts
@@ -1,4 +1,3 @@
 export * from './error-utils'
 export * from './requests'
 export * from './responses'
-export * from './domainState'