+++ packages/protocol/contracts/common/Accounts.sol
@@ -5,6 +5,7 @@ import "openzeppelin-solidity/contracts/ownership/Ownable.sol";
 
 import "./interfaces/IAccounts.sol";
 
+import "../common/FixidityLib.sol";
 import "../common/Initializable.sol";
 import "../common/interfaces/ICeloVersionedContract.sol";
 import "../common/Signatures.sol";
@@ -19,6 +20,7 @@ contract Accounts is
   Initializable,
   UsingRegistry
 {
+  using FixidityLib for FixidityLib.Fraction;
   using SafeMath for uint256;
 
   struct Signers {
@@ -57,6 +59,13 @@ contract Accounts is
     string metadataURL;
   }
 
+  struct PaymentDelegation {
+    // Address that should receive a fraction of validator payments.
+    address beneficiary;
+    // Fraction of payment to delegate to `beneficiary`.
+    FixidityLib.Fraction fraction;
+  }
+
   mapping(address => Account) internal accounts;
   // Maps authorized signers to the account that provided the authorization.
   mapping(address => address) public authorizedBy;
@@ -73,6 +82,9 @@ contract Accounts is
 
   // A per-account list of CIP8 storage roots, bypassing CIP3.
   mapping(address => bytes[]) public offchainStorageRoots;
+
+  // Optional per-account validator payment delegation information.
+  mapping(address => PaymentDelegation) internal paymentDelegations;
 
   bytes32 constant ValidatorSigner = keccak256(abi.encodePacked("celo.org/core/validator"));
   bytes32 constant AttestationSigner = keccak256(abi.encodePacked("celo.org/core/attestation"));
@@ -102,6 +114,7 @@ contract Accounts is
   event AccountCreated(address indexed account);
   event OffchainStorageRootAdded(address indexed account, bytes url);
   event OffchainStorageRootRemoved(address indexed account, bytes url, uint256 index);
+  event PaymentDelegationSet(address indexed beneficiary, uint256 fraction);
 
   /**
    * @notice Sets initialized == true on implementation contracts
@@ -297,6 +310,31 @@ contract Accounts is
     }
 
     return (concatenated, lengths);
+  }
+
+  /**
+   * @notice Sets validator payment delegation settings.
+   * @param beneficiary The address that should receive a portion of vaidator
+   * payments.
+   * @param fraction The fraction of the validator's payment that should be
+   * diverted to `beneficiary` every epoch, given as FixidyLib value. Must not
+   * be greater than 1.
+   */
+  function setPaymentDelegation(address beneficiary, uint256 fraction) public {
+    require(isAccount(msg.sender), "Not an account");
+    FixidityLib.Fraction memory f = FixidityLib.wrap(fraction);
+    require(f.lte(FixidityLib.fixed1()), "Fraction must not be greater than 1");
+    paymentDelegations[msg.sender] = PaymentDelegation(beneficiary, f);
+    emit PaymentDelegationSet(beneficiary, fraction);
+  }
+
+  /**
+   * @notice Gets validator payment delegation settings.
+   * @return Beneficiary address and fraction of payment delegated.
+   */
+  function getPaymentDelegation(address account) external view returns (address, uint256) {
+    PaymentDelegation storage delegation = paymentDelegations[account];
+    return (delegation.beneficiary, delegation.fraction.unwrap());
   }
 
   /**